Fix deploy file, get and create collections in SDK component
