+++ src/components/Header/ConnectButton.tsx
@@ -4,7 +4,7 @@ import { withRouter } from 'react-router-dom';
 import styled from 'styled-components'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSignIn } from '@fortawesome/free-solid-svg-icons';
-import { AppStateContext } from '../../App';
+import { AppStateContext } from 'SDK/WalletConnectSDK';
 import { ellipseAddress } from '../../helpers/utilities';
 import Button from '../Button';
