+++ src/views/Marketplace.tsx
@@ -1,47 +1,22 @@
-import React, { FC } from 'react'
-import styled, { keyframes } from 'styled-components'
+import React, { FC, useContext, useEffect } from 'react';
+import styled, { keyframes } from 'styled-components';
 import { Collections } from 'components';
-import { nftImage } from 'assets';
-
-const COLLECTIONS_DUMMY = [
-  {
-    id: 0,
-    image: nftImage,
-    smallImage: nftImage,
-    name: "New generation",
-    creator: "MisterPizza",
-    description: "asdcxzcxzczxczcxzczxcxzczxczcxzczcxxzcxzcxzccxzcz dsadsa"
-  },
-  {
-    id: 1,
-    image: nftImage,
-    smallImage: nftImage,
-    name: "New generation",
-    creator: "MisterPizza",
-    description: "asdcxzcxzczxczcxzczxcxzczxczcxzczcxxzcxzcxzccxzcz dsadsa"
-  },
-  {
-    id: 2,
-    image: nftImage,
-    smallImage: nftImage,
-    name: "New generation",
-    creator: "MisterPizza",
-    description: "asdcxzcxzczxczcxzczxcxzczxczcxzczcxxzcxzcxzccxzcz dsadsa"
-  },  {
-    id: 3,
-    image: nftImage,
-    smallImage: nftImage,
-    name: "New generation",
-    creator: "MisterPizza",
-    description: "asdcxzcxzczxczcxzczxcxzczxczcxzczcxxzcxzcxzccxzcz dsadsa"
-  }
-];
+import { AppStateContext } from 'SDK/WalletConnectSDK';
 
 const Marketplace: FC = () => {
+  const { state, getCollections } = useContext(AppStateContext);
+  const { collections, connected } = state;
+
+  useEffect(() => {
+    if (connected) {
+      getCollections();
+    }
+  }, [connected]);
+
   return (
     <MarketplaceWrapper>
       <MarketplaceTitle>{"Explore collections"}</MarketplaceTitle>
-      <Collections collections={COLLECTIONS_DUMMY}/>
+      <Collections collections={collections}/>
     </MarketplaceWrapper>
   );
 };
